backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
google-generativeai>=0.8.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
authlib==1.3.0
httpx>=0.28.1
itsdangerous==2.2.0
